Simulate: add rtc IO device

+++ sim/ioe/timer.cpp
@@ -22,7 +22,7 @@ uint64_t get_time() {
 }
 
 // ----- IO timer device -----
-static uint8_t *rtc_port_base = NULL;
+static uint32_t *rtc_port_base = NULL;
 
 static void rtc_io_handler(uint32_t offset, int len, bool is_write) {
     Assert(offset == 0 || offset == 4, "offset is %d\n", offset);
@@ -34,7 +34,7 @@ static void rtc_io_handler(uint32_t offset, int len, bool is_write) {
 }
 
 void init_timer() {
-    rtc_port_base = (uint8_t *)new_space(8);
+    rtc_port_base = (uint32_t *)new_space(8);
 
     add_mmio_map("rtc", CONFIG_RTC_MMIO, rtc_port_base, 8, rtc_io_handler);
 }
